requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
biopython==1.85
cattrs==24.1.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chembl-webresource-client==0.10.9
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
easydict==1.13
et_xmlfile==2.0.0
fastapi==0.115.12
fasteners==0.19
fonttools==4.56.0
GridDataFormats==1.0.2
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.1
markdown-it-py==3.0.0
matplotlib==3.10.0
mda-xdrlib==0.2.0
mdahole2==0.5.0
MDAnalysis==2.8.0
mdurl==0.1.2
mmtf-python==1.1.3
molecule-resolver==0.3.8
mrcfile==1.5.4
msgpack==1.1.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathsimanalysis==1.2.0
periodictable==2.0.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.52
PubChemPy==1.0.4
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyvips==3.0.0
rdflib==7.2.1
rdkit==2024.9.5
regex==2023.12.25
requests==2.32.3
requests-cache==1.2.1
rich==13.9.4
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
starlette==0.46.1
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
url-normalize==1.4.3
urllib3==2.3.0
uvicorn==0.34.0
waterdynamics==1.2.0
wcwidth==0.2.13
wheel==0.45.1
xmltodict==0.13.0
